requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
Pillow>=10.0,<12.0
pytest>=8.0
pytest-django>=4.8
pytest-cov==5.0.0